chore: install dev dependencies outside of docker
Dev containers already was mounting the project after building the dockerfile so the copy was duplicative. Additionally, the dockerfile install was installing the wrong set of files. The pip install needs to happen in post install so that it can act on the files that are mounted by the dev containers extension.